+++ pyqcrbox/pyqcrbox/sql_models/parameter_spec/parameter_spec.py
@@ -3,24 +3,25 @@ from typing import Annotated, Union
 
 from pydantic import Field, Tag, TypeAdapter
 
-# from .base_parameter_spec import SENTINEL_UNDEFINED
+from pyqcrbox import logger
+
 from .builtin_parameter_types import BoolParameterSpec, FloatParameterSpec, IntParameterSpec, StrParameterSpec
 from .filesystem_path_parameters import (
     CifDataFileParameterSpec,
     DataFileParameterSpec,
-    FolderPathParameterSpec,
-    GenericInputPathParameterSpec,
     GenericOutputPathParameterSpec,
-    InputCifParameterSpec,
-    InputFolderParameterSpec,
     OutputCifParameterSpec,
-    WorkCifParameterSpec,
+    # FolderPathParameterSpec,
+    # GenericInputPathParameterSpec,
+    # InputCifParameterSpec,
+    # InputFolderParameterSpec,
+    # WorkCifParameterSpec,
 )
 
 __all__ = ["ParameterSpecDiscriminatedUnion"]
 
 
-ParameterSpecTaggedUnion = Union[
+ParameterSpecTaggedUnion = Union[  # noqa: UP007
     #
     # Builtin types
     #
@@ -31,18 +32,21 @@ ParameterSpecTaggedUnion = Union[
     #
     # File/directory types with QCrBox-specific logic
     #
-    Annotated[InputCifParameterSpec, Tag("QCrBox.input_cif")],
-    Annotated[GenericInputPathParameterSpec, Tag("QCrBox.input_path")],
     Annotated[OutputCifParameterSpec, Tag("QCrBox.output_cif")],
-    Annotated[GenericOutputPathParameterSpec, Tag("QCrBox.output_path")],
-    Annotated[WorkCifParameterSpec, Tag("QCrBox.work_cif")],
-    Annotated[FolderPathParameterSpec, Tag("QCrBox.folder_path")],
     Annotated[DataFileParameterSpec, Tag("QCrBox.data_file")],
     Annotated[CifDataFileParameterSpec, Tag("QCrBox.cif_data_file")],
-    Annotated[InputFolderParameterSpec, Tag("QCrBox.input_folder")],
+    Annotated[GenericOutputPathParameterSpec, Tag("QCrBox.output_path")],
+    #
+    # Deprecated parameters
+    #
+    # Annotated[InputCifParameterSpec, Tag("QCrBox.input_cif")],
+    # Annotated[GenericInputPathParameterSpec, Tag("QCrBox.input_path")],
+    # Annotated[WorkCifParameterSpec, Tag("QCrBox.work_cif")],
+    # Annotated[FolderPathParameterSpec, Tag("QCrBox.folder_path")],
+    # Annotated[InputFolderParameterSpec, Tag("QCrBox.input_folder")],
 ]
-ParameterSpecDiscriminatedUnion = Annotated[ParameterSpecTaggedUnion, Field(discriminator="dtype")]
 
+ParameterSpecDiscriminatedUnion = Annotated[ParameterSpecTaggedUnion, Field(discriminator="dtype")]
 parameter_spec_adapter: TypeAdapter[ParameterSpecTaggedUnion] = TypeAdapter(ParameterSpecDiscriminatedUnion)
 
 
@@ -50,23 +54,20 @@ def get_param_spec_from_json(param_spec_json: dict) -> ParameterSpecDiscriminate
     return parameter_spec_adapter.validate_python(param_spec_json)
 
 
-def get_param_spec_from_signature_param(p: inspect.Parameter) -> ParameterSpecDiscriminatedUnion:
+def get_param_spec_from_signature_param(p: inspect.Parameter) -> ParameterSpecDiscriminatedUnion | None:
     if p.annotation == inspect._empty:
+        logger.error(f"Parameter {p.name} has no annotation")
         return None
 
     dtype = p.annotation.__name__
     is_required = p.default == inspect._empty
     default_value = None if is_required else repr(p.default)
 
-    param_spec_json = {"name": p.name, "dtype": dtype, "required": is_required, "default_value": default_value}
+    param_spec_json = {
+        "name": p.name,
+        "dtype": dtype,
+        "default_value": default_value,
+        "description": "dummy description which can be whatever for this validation step",
+        "required": True,
+    }
     return get_param_spec_from_json(param_spec_json)
-
-
-def ParameterSpec(**kwargs):
-    return get_param_spec_from_json(kwargs)
-    # if isinstance(data, dict):
-    #     return get_param_spec_from_json(data)
-    # elif isinstance(data, inspect.Parameter):
-    #     return get_param_spec_from_signature_param(data)
-    # else:
-    #     raise TypeError(f"Cannot instantiate ParameterSpec from input data: {data}")